Normalise casing for pg:copy
This commit normalises the casing for both config vars and
attachment names when performing a pg:copy. This is required because
it is possible for config vars and/or attachment names to be of
differing cases, resulting in value fetch against the `config` option
failing.

When the request does not contain the expected keys in JSON,
a 500 is returned from Transferatu without explaining the problem,
resulting in poor user experience. This commit attempts to address
one cause of missing `to_url` and `from_url` keys.